+++ laxar-mocks.js
@@ -48,20 +48,47 @@ export const TEST_WIDGET_ID = 'test-widget';
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+let fixtures = {};
+
 /**
- * This is only used by tooling such as the LaxarJS spec-loader.
- * Can be used to specify setup-fixtures for widget/activity tests.
+ * Allows to provide assets and configuration to be used for testing a widget.
  *
- * Spec-runners may add entries to this map to provision widget specs with options that will automatically be
- * picked up by `setupForWidget`. For example, the laxar-mocks spec-loader for webpack puts the `artifacts`,
- * `adapter` and `descriptor` options here.
+ * This should be used by tooling such as the LaxarJS spec-loader to specify spec fixtures in advance.
  *
  * Options passed by the spec-test to {@link #setupForWidget} will take precedence over these values.
+
+ * When using the spec-loader, something like the following code will be generated:
  *
- * @name fixtures
- * @type {Object}
+ * ```js
+ * ( require( 'laxar-mocks' ).init( {
+ *    descriptor: require( '../widget.json' ),
+ *    artifacts: require( 'laxar-loader?widget=example-widget' ),
+ *    adapter: require( 'laxar-' + fixtures.descriptor.integration.technology + '-adapter' )
+ * } );
+ * // ... spec test code ...
+ * ```
+ *
+ * @param {Object} specFixtures
+ *    modules and artifacts that are required for running the widget spec test
+ * @param {Object} specFixtures.descriptor
+ *    the widget's JSON descriptor, containing name, integration and feature validation schema
+ * @param {Object} specFixtures.artifacts
+ *    an artifacts listing containing modules and assets for the widget under test and its required controls.
+ *    Usually generated using the `laxar-tooling` library, or the `laxar-loader` for webpack
+ * @param {Object} [specFixtures.adapter]
+ *    an adapter module for the integration technology of the widget under test. Omit for "plain" widgets
  */
-export const fixtures = {};
+export function init( specFixtures ) {
+   assert( specFixtures.descriptor ).hasType( Object ).isNotNull();
+   assert( specFixtures.artifacts ).hasType( Object ).isNotNull();
+   assert( specFixtures.artifacts.widgets ).hasType( Array ).isNotNull();
+   assert( specFixtures.artifacts.aliases ).hasType( Object ).isNotNull();
+   if( specFixtures.adapter ) {
+      assert( specFixtures.adapter.technology ).hasType( String ).isNotNull();
+      assert( specFixtures.adapter.bootstrap ).hasType( Function ).isNotNull();
+   }
+   fixtures = specFixtures;
+}
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -451,14 +478,11 @@ export function triggerStartupEvents( optionalEvents = {} ) {
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /**
- * Creates the setup function for a widget test, using externally provided fixtures.
+ * Creates the setup function for a widget test, using fixtures that were provided through {@link #init}.
  *
- * This is the recommended way to setup your widget test. For this to work, *this* module's `fixtures` export
- * needs to be initialized with the following properties:
- *
- *   - `descriptor` - the widget's JSON descriptor,
- *   - `adapter` - the adapter module for the widget's integration technology (use `null` for "plain"),
- *   - `artifacts` - an artifact listing containing the assets of the widget and its controls.
+ * This is the recommended way to setup your widget test. For this to work without manully providing options,
+ * this module's `init` method must must have been called already, providing `descriptor`, `adapter` and
+ * `artifacts`.
  *
  * When webpack loads spec-tests through the `laxar-mocks/spec-loader`, fixtures are provided automatically.
  * To manually provide these fixtures, controlling every aspect of your test environment, pass them using the
@@ -479,20 +503,6 @@ export function triggerStartupEvents( optionalEvents = {} ) {
  * } );
  * ```
  *
- * When using the spec-loader, something like the following code will be generated:
- *
- * ```js
- * ( fixtures => {
- *    fixtures.descriptor = require( '../widget.json' );
- *    fixtures.artifacts = require( 'laxar-loader?widget=example-widget' );
- *    fixtures.adapter = require( 'laxar-' + fixtures.descriptor.integration.technology + '-adapter' );
- * } )( require( 'laxar-mocks' ).fixtures );
- * import * as axMocks from 'laxar-mocks';
- *
- * describe( 'An ExampleWidget', () => {
- *    // ... same as above ...
- * } );
- * ```
  * @param {Object} [optionalOptions]
  *    optional map of options
  * @param {Object} [optionalOptions.adapter=undefined]